Split internal objects file to simplificate code
