+++ src/business_logic/transaction/mod.rs
@@ -1,6 +1,5 @@
 pub mod error;
 pub mod fee;
-pub mod internal_objects;
 pub(crate) mod objects;
 pub mod state_objects;
 pub mod transactions;
