+++ src/business_logic/transaction/objects/mod.rs
@@ -1,1 +1,3 @@
+pub(crate) mod interal_declare;
+pub(crate) mod internal_deploy;
 pub(crate) mod internal_invoke_function;
