+++ src/testing/starknet_state_error.rs
@@ -1,7 +1,9 @@
 use thiserror::Error;
 
 use crate::{
-    business_logic::execution::execution_errors::ExecutionError,
+    business_logic::{
+        execution::execution_errors::ExecutionError, transaction::error::TransactionError,
+    },
     core::errors::{state_errors::StateError, syscall_handler_errors::SyscallHandlerError},
 };
 
@@ -15,4 +17,6 @@ pub enum StarknetStateError {
     StateException(#[from] StateError),
     #[error(transparent)]
     ExecuteException(#[from] ExecutionError),
+    #[error(transparent)]
+    TransactionException(#[from] TransactionError),
 }
